Move Color, Privacy & Type from project/Creates to project/edit
