+++ frontend/src/components/project/Creates/ProjectCreate.jsx
@@ -4,9 +4,9 @@ import { useMutation } from "@tanstack/react-query"
 import { useTheme } from "styled-components"
 
 import { useModalWindowCloseContext } from "@components/common/ModalWindow"
-import Color from "@components/project/Creates/Color"
-import Privacy from "@components/project/Creates/Privacy"
-import Type from "@components/project/Creates/Type"
+import Color from "@components/project/edit/Color"
+import Privacy from "@components/project/edit/Privacy"
+import Type from "@components/project/edit/Type"
 import { getProjectColor } from "@components/project/Creates/palettes"
 import Middle from "@components/project/edit/Middle"
 import TitleInput from "@components/project/edit/TitleInput"
